Modified Dictionary Creation.
when user enters the id, the program first searches for a txt file. If the file is found, the dictionary is created using this file. Otherwise, it makes an API call using the given id, retrieves the data, and if they lead to a valid dictionary, it creates a txt file.